python/chatgpt/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mypy-extensions==1.0.0
openai==1.29.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pydantic==2.7.1
pydantic-core==2.18.2
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
termcolor==2.4.0
tomli==2.0.1
tqdm==4.66.4
typing-extensions==4.11.0
urllib3==2.2.1
